m: create base class for fixture
